task: refactor
